Fix bugs of int overflow

+++ internal/ceres/block_random_access_diagonal_matrix_test.cc
@@ -99,7 +99,7 @@ class BlockRandomAccessDiagonalMatrixTest : public ::testing::Test {
 
  protected:
   ContextImpl context_;
-  int num_nonzeros_;
+  int64_t num_nonzeros_;
   std::unique_ptr<BlockRandomAccessDiagonalMatrix> m_;
 };
