+++ internal/ceres/dynamic_compressed_row_sparse_matrix.cc
@@ -35,7 +35,7 @@
 namespace ceres::internal {
 
 DynamicCompressedRowSparseMatrix::DynamicCompressedRowSparseMatrix(
-    int num_rows, int num_cols, int initial_max_num_nonzeros)
+    int num_rows, int num_cols, int64_t initial_max_num_nonzeros)
     : CompressedRowSparseMatrix(num_rows, num_cols, initial_max_num_nonzeros) {
   dynamic_cols_.resize(num_rows);
   dynamic_values_.resize(num_rows);
